add preserve_dir_tree option to has_attached_file and skip tree deletion if true

+++ lib/paperclip/storage/filesystem.rb
@@ -65,17 +65,19 @@ module Paperclip
           rescue Errno::ENOENT => e
             # ignore file-not-found, let everything else pass
           end
-          begin
-            while(true)
-              path = File.dirname(path)
-              FileUtils.rmdir(path)
-              break if File.exist?(path) # Ruby 1.9.2 does not raise if the removal failed.
-            end
-          rescue Errno::EEXIST, Errno::ENOTEMPTY, Errno::ENOENT, Errno::EINVAL, Errno::ENOTDIR, Errno::EACCES
+          if @options[:preserve_dir_tree] == false 
+            begin
+              while(true)
+               path = File.dirname(path)
+               FileUtils.rmdir(path)
+               break if File.exist?(path) # Ruby 1.9.2 does not raise if the removal failed.
+              end
+            rescue Errno::EEXIST, Errno::ENOTEMPTY, Errno::ENOENT, Errno::EINVAL, Errno::ENOTDIR, Errno::EACCES
             # Stop trying to remove parent directories
-          rescue SystemCallError => e
+            rescue SystemCallError => e
             log("There was an unexpected error while deleting directories: #{e.class}")
             # Ignore it
+            end
           end
         end
         @queued_for_delete = []
